pyproject.toml
[project]
name = "veto-budget-mcp"
version = "1.0.0"
description = "Veto Budget Agent - MCP server for budget management and AI-powered financial decisions"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp",
    "supabase>=2.0.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
